Completely re-written setup so that it is even simpler.
This re-write is the start of a project wide re-write of how SymSteam recognises drives. SymSteam now stores the UUID of a drive and requires that the drive is a HFS+ drive (since it needs a UUID). In the future, the Steam drive will be recognised based on its UUID and not the drive's name which will fix a lot of potential issues.

N.B. At this build (build 2, since I've just decided to start using build numbers), SymSteam doesn't work since I haven't implemented the code to handle detecting drives based on their UUID.