Re-implemented drive detection to use the Disk Arbitration framework.
SymSteam now uses the Core Foundation callbacks provided by DADisk rather than the notifications provided by NSWorkspace for device mounts/unmounts. These callbacks are far more useful than the ones provided by NSWorkspace's notifications as they allow us to gain access to a DADiskRef when a drive is unplugged or unmounted.

There's a little bit of trickery going on in this new implementation. Since DADisk doesn't provide callbacks for drive mounts/un-mounts, we have to use the mount/un-mount approval callbacks to do our work instead. These callbacks only appear if a drive is ejected though so in order to get callbacks for when a drive is suddenly unplugged (without being ejected), we also register for a drive disappearance callback.