Refactored SCSetupController a bit more.
Added 2 new methods. One to extract the path to an external drive from the path to a folder on that drive and one to verify that the disk is a valid steam drive (it just wraps two other methods in the class into one).

In addition, the symbolic link creation method will now check to see if the SteamDrive is connected in order to avoid creating two symbolic SteamApps folders.