Implemented error notification functionality.
This is early stuff and highly untested. I've added a category to SCUserNotificationCenter that displays a notification as normal and attaches an NSError object to the userInfo property. When the notification is clicked, the NSError is reconstructed from the userInfo dictionary and is then displayed to the user via a NSAlert. It works though I haven't used it anywhere in the application yet.